@@ -170,7 +170,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions
             }
         }
 
-        private IReadOnlyCollection<string> GetProducedResources()
+        private List<string> GetProducedResources()
         {
             var producedResources = new List<string>();
 
