@@ -93,9 +93,9 @@ namespace Microsoft.Health.Fhir.Api.UnitTests.Features.Headers
         [Fact]
         public void WhenAddingSameHeaderTwice_ThenOnlyOneHeaderIsPresent()
         {
-            Assert.Throws<ArgumentException>(() => FhirResult.Create(_mockResource)
+            FhirResult.Create(_mockResource)
                 .SetLastModifiedHeader()
-                .SetLastModifiedHeader());
+                .SetLastModifiedHeader();
         }
 
         [Fact]
