@@ -358,7 +358,7 @@ namespace Microsoft.Health.Internal.Fhir.RegisterAndMonitorImport
             else
             {
                 value = value.Trim();
-                if (value.StartsWith("{", StringComparison.Ordinal) && value.EndsWith("}", StringComparison.Ordinal))
+                if (value.StartsWith('{') && value.EndsWith('}'))
                 {
                     try
                     {
